refactor(queries): split the action log from the activation worklist
## Summary

`sam/queries/xras_actions.py` was 906 lines fusing two concerns over two tables
with no shared helpers across the seam:

| lines | concern | table |
|---|---|---|
| ~610 | action-log read model: vocabulary, filters, pagination, facets | `xras_action_log` |
| ~296 | activation worklist: derived operator state, timeline, recipients | `xras_activation_event` |

The forms layer already splits exactly this way — `schemas/forms/xras.py` and
`xras_activation.py` — so the query layer was inconsistent with its own sibling.

Now `queries/xras_actions.py` (610) and `queries/xras_activation.py` (~340).

The one genuinely shared thing is the `projcode_result` OR `request_number` join,
which `xras_activation` imports as `action_names_project` + `_LATEST_ACTION_ORDER`
rather than re-spelling. That is stated at the top of the new module with a
pointer to why: re-spelling it is what let the pending card and the provenance
stamp disagree on a same-second tie two commits ago. The split makes that
dependency explicit where it used to be accidental co-location.

Also dropped `EmailAddress` / `User` from the moved code's imports — it reaches
its people through `project.lead` / `project.admin`, so neither was ever used.

### Call sites

All five names are still re-exported from the `sam.queries` facade, so nothing
that goes through it changed. The two modules that import the submodule directly
— the allocations blueprint and this module's own tests — were repointed.

### Test Results

`pytest -q` → **5,246 passed**, 36 skipped, 1 xfailed — unchanged from the
previous commit, which is the point. Leak check 0.
