+++ src/sam/queries/__init__.py
@@ -116,15 +116,17 @@ from .xras_actions import (
     XRAS_REQUEST_TOKEN_PREFIXES,
     canonical_action_type,
     count_recent_xras_actions,
-    count_xras_dismissed_pending,
-    get_latest_xras_action_id,
     get_observed_action_types,
     get_projects_by_ids,
     get_recent_xras_actions,
+    summarize_xras_actions,
+)
+from .xras_activation import (
+    count_xras_dismissed_pending,
+    get_latest_xras_action_id,
     get_xras_activation_events,
     get_xras_pending_activation,
     get_xras_pending_recipients,
-    summarize_xras_actions,
 )
 
 # Usage cache
